backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
motor==3.3.1
bcrypt==4.0.1
pandas==2.1.1
numpy==1.26.1
scikit-learn==1.3.2
deap==1.4.1
matplotlib==3.8.1
python-dateutil==2.8.2 